# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ДоВуз-Портал API
  version: 0.1.0
  description: |
    Контракт бэкенда АИС «ДоВуз-Портал». Снят со слоя `frontend/src/app/api/`
    и повторяет его как есть, включая смешанное именование полей: часть ответов
    в snake_case, часть в camelCase, query-параметры в camelCase.

    Аутентификация — HTTP-only кука сессии, которую выставляют `/auth/login`
    и `/auth/register`. Ошибки всегда приходят в формате `{"message": "..."}`.
    401 — «не вошёл» (фронт разлогинивает), 403 — «вошёл, но нельзя».
servers:
  - url: /v1
    description: Относительно хоста, с которого отдана спека; локально — http://localhost:8080/v1
security:
  - sessionCookie: []
...
paths:
...
  /roles:
    get:
      tags: [roles]
      operationId: getRoles
      summary: Все роли
      responses:
        '200':
          description: Роли
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoleResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /subject-topics:
    get:
      tags: [subject-topics]
      operationId: getSubjectTopics
      summary: Темы предмета
      parameters:
        - name: subjectId
          in: query
          required: true
          schema:
            type: integer
        - $ref: '#/components/parameters/PageNumber'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Страница тем
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectTopicPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [subject-topics]
      operationId: createSubjectTopic
      summary: Создать тему
      description: Администратору и модераторам, назначенным на предмет.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectTopicCreateRequest'
      responses:
        '201':
          description: Тема создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectTopicResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /subject-topics/{id}:
    get:
      tags: [subject-topics]
      operationId: getSubjectTopic
      summary: Тема по id
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Тема
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectTopicResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [subject-topics]
      operationId: updateSubjectTopic
      summary: Переименовать тему
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NameRequest'
      responses:
        '200':
          description: Тема
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectTopicResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [subject-topics]
      operationId: deleteSubjectTopic
      summary: Удалить тему
      description: Вместе с публикациями и их файлами.
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Удалена
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /publications:
    get:
      tags: [publications]
      operationId: getPublicationsByTopic
      summary: Публикации темы
      parameters:
        - name: subjectTopicId
          in: query
          required: true
          schema:
            type: integer
        - $ref: '#/components/parameters/PageNumber'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Страница публикаций, новые сверху
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicationTitlePage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [publications]
      operationId: createPublication
      summary: Создать публикацию в теме
      description: Администратору и модераторам, назначенным на предмет темы.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [title, subjectTopicId]
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 300
                description:
                  type: string
                  maxLength: 20000
                subjectTopicId:
                  type: integer
                files:
                  type: array
                  description: До 10 файлов, каждый до 50 МБ
                  maxItems: 10
                  items:
                    type: string
                    format: binary
      responses:
        '201':
          description: Публикация создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicationDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
  /publications/{id}:
    get:
      tags: [publications]
      operationId: getPublication
      summary: Публикация по id
      description: Новости открыты всем, публикации из тем — только после входа.
      security:
        - {}
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Публикация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicationDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [publications]
      operationId: deletePublication
      summary: Удалить публикацию
      description: |
        Публикацию из темы — администратор или модератор предмета,
        новость — администратор или её автор.
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Удалена
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /schools:
    get:
      tags: [schools]
      operationId: getSchools
      summary: Все школы
      responses:
        '200':
          description: Школы
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SchoolResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [schools]
      operationId: createSchool
      summary: Создать школу
      description: Только для администратора. Название уникально без учёта регистра.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NameRequest'
      responses:
        '201':
          description: Школа создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchoolResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /schools/{id}:
    delete:
      tags: [schools]
      operationId: deleteSchool
      summary: Удалить школу
      description: Вместе с классами. Ученики остаются, но теряют привязку к школе и классу.
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Удалена
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: dovuz_session
...
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
    PageNumber:
      name: pageNumber
      in: query
      description: Номер страницы, с нуля
      schema:
        type: integer
        minimum: 0
        maximum: 1000000
        default: 0
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 10
...
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Нужно войти в систему
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Недостаточно прав
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Не найдено
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    PayloadTooLarge:
      description: Файл слишком большой
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
    RoleTitle:
      type: string
      description: Роль так, как её показывает и сравнивает фронт
      enum: [Администратор, Модератор, Пользователь]
      x-enum-varnames: [ADMIN, MODERATOR, USER]
    RoleCode:
      type: string
      enum: [ROLE_ADMIN, ROLE_MODERATOR, ROLE_USER]
...
    Name:
      type: string
      minLength: 1
      maxLength: 300
    PageMeta:
      type: object
      required: [total_pages, total_size, page_number, page_size]
      properties:
        total_pages:
          type: integer
        total_size:
          type: integer
        page_number:
          type: integer
        page_size:
          type: integer
...
    RoleResponse:
      type: object
      description: '`label`, `role` и `name` — одно и то же русское название: фронт читает все три'
      required: [id, label, role, name, code]
      properties:
        id:
          type: integer
        label:
          $ref: '#/components/schemas/RoleTitle'
        role:
          $ref: '#/components/schemas/RoleTitle'
        name:
          $ref: '#/components/schemas/RoleTitle'
        code:
          $ref: '#/components/schemas/RoleCode'
    NameRequest:
      type: object
      required: [name]
      properties:
        name:
          $ref: '#/components/schemas/Name'
...
    SubjectRef:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
    SubjectTopicResponse:
      type: object
      required: [id, name, subject]
      properties:
        id:
          type: integer
        name:
          type: string
        subject:
          $ref: '#/components/schemas/SubjectRef'
    SubjectTopicCreateRequest:
      type: object
      required: [name, subjectId]
      properties:
        name:
          $ref: '#/components/schemas/Name'
        subjectId:
          type: integer
    SubjectTopicPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [content]
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/SubjectTopicResponse'
    FileInfo:
      type: object
      required: [id, file_name_in_directory, initial_file_name]
      properties:
        id:
          type: integer
        file_name_in_directory:
          type: string
          description: Имя в хранилище, по нему файл скачивается
        initial_file_name:
          type: string
          description: Имя, под которым файл загрузили
...
    PublicationDetailResponse:
      type: object
      required: [id, title, description, created_at, nickname, files, subject_topic_id]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
        nickname:
          type: [string, 'null']
          description: Автор; `null`, если его учётную запись удалили
        files:
          type: array
          items:
            $ref: '#/components/schemas/FileInfo'
        subject_topic_id:
          type: [integer, 'null']
          description: Тема; у новостей `null`
    PublicationTitleAndId:
      type: object
      required: [id, title]
      properties:
        id:
          type: integer
        title:
          type: string
    PublicationTitlePage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [content]
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/PublicationTitleAndId'
...
    SchoolResponse:
      type: object
      required: [id, name, topic_deadline]
      properties:
        id:
          type: integer
        name:
          type: string
        topic_deadline:
          type: [string, 'null']
          format: date-time
          description: Дедлайн выбора темы проекта
...